simple script to debug the judging: try_to_judge_single_submission.php

+++ lib/Tempdir.php
@@ -6,12 +6,16 @@
 
 class Tempdir {
 	public $dir;
+	private $delete;
 	
 	function __construct($in,$name) {
 		$this->dir = SystemUtil::temporary_directory($in,$name);
+		$this->delete = true;
 	}
 	function __destruct() {
-		SystemUtil::delete_directory($this->dir);
+		if ($this->delete) {
+			SystemUtil::delete_directory($this->dir);
+		}
 	}
 	
 	function file($name) {
@@ -19,5 +23,9 @@ class Tempdir {
 	}
 	function create_parent_dirs($name) {
 		@mkdir(dirname($this->file($name)), 0777, true);
+	}
+	function get_and_keep() {
+		$this->delete = false;
+		return $this->dir;
 	}
 }
